+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/DemonHunter/DeathSpeakerBlackthorn.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.DemonHunter;
+
+public class DeathSpeakerBlackthorn : ICardWithHighlight
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Demonhunter.DeathSpeakerBlackthorn;
+
+	public HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck) =>
+		HighlightColorHelper.GetHighlightColor(card.HasDeathrattle() && card.Cost <= 5);
+}