+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/DemonHunter/DoubleJump.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+using HearthDb.Enums;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.DemonHunter;
+
+public class DoubleJump : ICardWithHighlight
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Demonhunter.DoubleJump;
+
+	public HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck) =>
+		HighlightColorHelper.GetHighlightColor(
+			card.GetTag(GameTag.OUTCAST) > 0);
+}