+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Neutral/CrushclawEnforcer.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Neutral;
+
+public class CrushclawEnforcer : ICardWithHighlight
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Neutral.CrushclawEnforcer;
+
+	public HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck) =>
+		HighlightColorHelper.GetHighlightColor(card.IsNaga());
+}