+++ Hearthstone Deck Tracker/Hearthstone/RelatedCardsSystem/Cards/Warlock/ArchwitchWillow.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+
+namespace Hearthstone_Deck_Tracker.Hearthstone.RelatedCardsSystem.Cards.Warlock;
+
+public class ArchwitchWillow : ICardWithHighlight
+{
+	public string GetCardId() => HearthDb.CardIds.Collectible.Warlock.ArchwitchWillow;
+
+	public HighlightColor ShouldHighlight(Card card, IEnumerable<Card> deck) =>
+		HighlightColorHelper.GetHighlightColor(card.IsDemon());
+}